+++ src/utility/helpers.js
@@ -1,4 +1,5 @@
 const assert = require('node:assert');
+const uuid = require('uuid');
 
 
 function IsType(type) {
@@ -25,6 +26,13 @@ function IsDerivedFrom(derived, base) {
   return IsTypeOf(derived, base) || derived instanceof base;
 }
 
+function IsValidUUID(idToTest) {
+  if (!IsTypeOf(idToTest, String)) {
+    return false;
+  }
+  return uuid.validate(idToTest) && uuid.version(idToTest) === 4;
+}
+
 function EnforceIsTypeOf(left, right) {
   assert(IsTypeOf(left, right), `Expected '${GetTypeName(left)}' to be of type '${GetTypeName(right)}'.`);
 }
@@ -33,6 +41,11 @@ function EnforceIsDerivedFrom(derived, base) {
   assert(IsDerivedFrom(derived, base), `Expected '${GetTypeName(derived)}' to be an instance of '${GetTypeName(base)}'.`);
 }
 
+function EnforceIsValidID(idToTest) {
+  EnforceIsTypeOf(idToTest, String);
+  assert(uuid.validate(idToTest), `"${idToTest}" is not in uuid format.`);
+  assert(uuid.version(idToTest) === 4, `"${idToTest}"'s uuid version is not 4. Version number is ${uuid.version(idToTest)} instead.`);
+}
 
 // Module Exports --------------------------------------------------------------
 module.exports = {
@@ -40,7 +53,9 @@ module.exports = {
   GetTypeName,
   IsTypeOf,
   IsDerivedFrom,
+  IsValidUUID,
   EnforceIsTypeOf,
-  EnforceIsDerivedFrom
+  EnforceIsDerivedFrom,
+  EnforceIsValidID
 };
 // Module Exports (END) --------------------------------------------------------